hex-opbatch: cleanup naming and headers for opbatch and related descriptors
