+++ ggml/src/ggml-hexagon/htp/ssm-conv.c
@@ -16,7 +16,7 @@
 #include "ggml-common.h"
 #include "htp-ctx.h"
 #include "hex-dma.h"
-#include "htp-msg.h"
+#include "htp-ops.h"
 #include "htp-ops.h"
 #include "hvx-utils.h"
